Stuff moved from PI to controller. Error when included executeControl in adc fixed.
