+++ Software/Development/SemesterProject/App/Include/adcMonitor.h
@@ -13,18 +13,16 @@
 //
 #include "DSP28x_Project.h"     // Device Headerfile and Examples Include File
 #include "analogAcquisitionManager.h"
+#include "errorManager.h"
 
-typedef enum
-{
-    NO_ERROR,
-    ERROR_HAS_HAPPENED,
-    IDLE
-} errorStatus;
 //
 // Function prototyping
 //
 errorStatus areAdcMeasurementsWithinThresholds(void);
 errorStatus areBatteryMeasurementsWithinThresholds(void);
 errorStatus areCurrentMeasurementsWithinThresholds(void);
+
+errorStatus getErrorStatusInBit(Uint16);
+
 
 #endif /* APP_INCLUDE_ADCMONITOR_H_ */
