+++ Software/Development/SemesterProject/App/PIController.c
@@ -16,16 +16,6 @@
 //
 #include "PIController.h"
 
-//
-// Quasi-global variables definition
-//
-static struct
-{
-    PIobject IqController;
-    PIobject IdController;
-    PIobject SpeedController;
-} PIControllerList;
-
 
 void PIObject_Constructor(PIobject *PIcontroller, float KP, float KI,
                           float saturationThreshold, int antiWindupFlag)
@@ -39,13 +29,6 @@ void PIObject_Constructor(PIobject *PIcontroller, float KP, float KI,
     PIcontroller->previousLimitedOutput = 0;
 
     PIcontroller->antiWindupFlag = antiWindupFlag;
-}
-
-void initPIControllers(void)
-{
-    PIObject_Constructor(&PIControllerList.IdController, KP_ID, KI_ID, CURRENT_LIMIT, 0);
-    PIObject_Constructor(&PIControllerList.IqController, KP_IQ, KI_IQ, CURRENT_LIMIT, 0);
-    PIObject_Constructor(&PIControllerList.SpeedController, KP_SPEED, KI_SPEED, SPEED_LIMIT, 1);
 }
 
 /*
@@ -89,23 +72,6 @@ float PiCalculation(PIobject *PIcontroller, float reference, float measuredValue
     return PIoutput;
 }
 
-/*
- * Interface functions to use PI controllers
- */
-float PiCalcualtionIQ(float reference, float measuredValue)
-{
-    return PiCalculation(&PIControllerList.IqController, reference, measuredValue);
-}
-
-float PiCalculationID(float reference, float measuredValue)
-{
-    return PiCalculation(&PIControllerList.IdController, reference, measuredValue);
-}
-
-float PiCalculationSpeed(float reference, float measuredValue)
-{
-    return PiCalculation(&PIControllerList.SpeedController, reference, measuredValue);
-}
 
 /*
  * Functions to check for saturation
